MissingNumber
You are given an integer array containing 1 to n but one of the numbers from 1 to n in the array is missing. You need to provide optimum solution to find the missing number. Number cannot be repeated in the array. Hints: You can use the following formula to get sum of 1 to n: n*(n+1)/2  Example:Array1= {7,5,6,1,4,2}; Missing number: 3;  Array2= {5,3,1,2};Missing number: 4   (answer:in web executable link or source code in github or paste your code in answer text field)